+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 
 - shrinked all images
 - made all tests (also agents) self contained
-
+- SDK: added IotTimeSeriesAggregateClient to the SDK
+- SDK: added SpectrumAnalysisClient to the SDK
+- CLI Command: mc spectrum-analysis: performs spectrum analysis on a sound file
 
 ## 3.5.3 - (Venetian Red Vienna) - May 2019 - Recovery